[SYCL] Split group and nd_item query-only core headers
This change reduces compile-time overhead in SYCL headers by splitting
`group.hpp` and `nd_item.hpp` into query-only core headers plus heavier
extra layers, and by introducing `access_base.hpp` so lightweight query
paths no longer pull in the full `access/access.hpp` umbrella.

- Add `detail/group_core.hpp` for the `group` class definition and query-only API.
- Add `detail/group_extra.hpp` for heavier functionality:
  - `private_memory`
  - `parallel_for_work_item`
  - async work-group copy helpers and definitions
- Convert `sycl/group.hpp` into a thin public aggregator.

- Add `detail/nd_item_core.hpp` for the `nd_item` class definition and query-only API.
- Add `detail/nd_item_extra.hpp` for heavier functionality:
  - `get_offset()`
  - `get_nd_range()`
  - async work-group copy definitions
  - wait helpers
  - root-group helpers
- Convert `sycl/nd_item.hpp` into a thin public aggregator.

- Add `sycl/access/access_base.hpp` as a lightweight home for:
  - access enums
  - `remove_decoration` utilities
- Keep `sycl/access/access.hpp` as the heavier umbrella.
- Retarget lightweight dependency sites to `access_base.hpp`:
  - `detail/fwd/multi_ptr.hpp`
  - `detail/spirv_memory_semantics.hpp`
  - query-only core headers

- Replace the public `nd_item.hpp` dependency with:
  - `detail/nd_item_core.hpp`
  - `detail/sub_group_core.hpp`
  - `detail/builder.hpp`
- Keep behavior unchanged while reducing the transitive include graph.

- Update `detail/builder.hpp` to match the new stateless `nd_item` core path.
- Adjust subgroup-related layering fallout needed by the new header closure.
- Regenerate include-deps tests and update the `layout_array.cpp` ABI baseline.

Measured with `count_trace_includes.py` against the `#21773` baseline.

- Transitive headers: 79 → 67
- SYCL headers benchmarked: 32 → 19
- stdlib headers benchmarked: 10 → 10

| Header                              | Before | After  | Delta            |
|-------------------------------------|--------|--------|------------------|
| free_function_queries.hpp (whole)   | 73.0 ms| 57.7 ms| -15.3 ms (-21%)  |
| sycl/nd_item.hpp                    | 72.0 ms| n/a    | eliminated       |
| sycl/group.hpp                      | 61.6 ms| n/a    | eliminated       |
| sycl/detail/helpers.hpp             | 32.2 ms| n/a    | eliminated       |
| sycl/detail/async_work_group_copy_ptr.hpp | 27.7 ms | n/a | eliminated |
| sycl/pointers.hpp                   | 24.3 ms| n/a    | eliminated       |
| sycl/access/access.hpp              | 24.4 ms| n/a    | eliminated       |
| sycl/detail/nd_item_core.hpp        | n/a    | 55.4 ms| new              |
| sycl/detail/group_core.hpp          | n/a    | 53.5 ms| new              |
| sycl/access/access_base.hpp         | n/a    | 10.4 ms| new              |

- `sycl/nd_item.hpp` → replaced by `detail/nd_item_core.hpp`
- `sycl/group.hpp` transitively disappears from the query-only path
- `sycl/pointers.hpp`
- `sycl/device_event.hpp`
- `sycl/nd_range.hpp`
- `sycl/detail/async_work_group_copy_ptr.hpp`
- `sycl/access/access.hpp`

- No functional changes intended.
- Public `sycl/group.hpp` and `sycl/nd_item.hpp` remain source-compatible.
- The remaining visible fallout is limited to header layering and test baseline updates.
- The `layout_array.cpp` update reflects record-layout dump output for the now-empty `nd_item` core type, not an intended runtime ABI change.

Depends on #21773
