+++ sycl/include/sycl/access/access.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include <sycl/detail/defines_elementary.hpp> // for __SYCL2020_DEPRECATED
+#include <sycl/access/access_base.hpp>
 
 #ifdef __SYCL_DEVICE_ONLY__
 #include <type_traits>
@@ -16,50 +16,6 @@
 
 namespace sycl {
 inline namespace _V1 {
-namespace access {
-
-enum class target {
-  device = 2014,
-  global_buffer __SYCL2020_DEPRECATED("use 'target::device' instead") = device,
-  constant_buffer __SYCL2020_DEPRECATED("use 'target::device' instead") = 2015,
-  local __SYCL2020_DEPRECATED("use `local_accessor` instead") = 2016,
-  image = 2017,
-  host_buffer __SYCL2020_DEPRECATED("use 'host_accessor' instead") = 2018,
-  host_image = 2019,
-  image_array = 2020,
-  host_task = 2021,
-};
-
-enum class mode {
-  read = 1024,
-  write = 1025,
-  read_write = 1026,
-  discard_write = 1027,
-  discard_read_write = 1028,
-  atomic = 1029
-};
-
-enum class fence_space {
-  local_space = 0,
-  global_space = 1,
-  global_and_local = 2
-};
-
-enum class placeholder { false_t = 0, true_t = 1 };
-
-enum class address_space : int {
-  private_space = 0,
-  global_space = 1,
-  constant_space __SYCL2020_DEPRECATED("sycl::access::address_space::constant_"
-                                       "space is deprecated since SYCL 2020") =
-      2,
-  local_space = 3,
-  generic_space = 4, // TODO generic_space address space is not supported yet
-};
-
-enum class decorated : int { no = 0, yes = 1, legacy = 2 };
-} // namespace access
-
 using access::target;
 using access_mode = access::mode;
 
@@ -207,61 +163,7 @@ struct deduce_AS
           std::remove_pointer_t<std::remove_reference_t<std::remove_cv_t<T>>>> {
 };
 #endif
-
-template <typename T> struct remove_decoration_impl {
-  using type = T;
-};
-
-#ifdef __SYCL_DEVICE_ONLY__
-template <typename T> struct remove_decoration_impl<__OPENCL_GLOBAL_AS__ T> {
-  using type = T;
-};
-
-template <typename T> struct remove_decoration_impl<__OPENCL_PRIVATE_AS__ T> {
-  using type = T;
-};
-
-template <typename T> struct remove_decoration_impl<__OPENCL_LOCAL_AS__ T> {
-  using type = T;
-};
-
-template <typename T> struct remove_decoration_impl<__OPENCL_CONSTANT_AS__ T> {
-  using type = T;
-};
-#endif // __SYCL_DEVICE_ONLY__
 } // namespace detail
-
-template <typename T> struct remove_decoration {
-  using type = typename detail::remove_decoration_impl<T>::type;
-};
-
-// Propagate through const qualifier.
-template <typename T> struct remove_decoration<const T> {
-  using type = const typename remove_decoration<T>::type;
-};
-
-// Propagate through pointer.
-template <typename T> struct remove_decoration<T *> {
-  using type = typename remove_decoration<T>::type *;
-};
-
-// Propagate through const qualified pointer.
-template <typename T> struct remove_decoration<const T *> {
-  using type = const typename remove_decoration<T>::type *;
-};
-
-// Propagate through reference.
-template <typename T> struct remove_decoration<T &> {
-  using type = typename remove_decoration<T>::type &;
-};
-
-// Propagate through const qualified reference.
-template <typename T> struct remove_decoration<const T &> {
-  using type = const typename remove_decoration<T>::type &;
-};
-
-template <typename T>
-using remove_decoration_t = typename remove_decoration<T>::type;
 
 #undef __OPENCL_GLOBAL_AS__
 #undef __OPENCL_LOCAL_AS__
