+++ sycl/include/sycl/detail/spirv_memory_semantics.hpp
@@ -8,9 +8,9 @@
 
 #pragma once
 
-#include <sycl/__spirv/spirv_types.hpp> // for MemorySemanticsMask
-#include <sycl/access/access.hpp>       // for fence_space
-#include <sycl/memory_enums.hpp>        // for memory_order
+#include <sycl/__spirv/spirv_types.hpp>
+#include <sycl/access/access_base.hpp>
+#include <sycl/memory_enums.hpp>
 
 #include <stdint.h>
 
@@ -27,31 +27,6 @@ inline constexpr uint32_t
 getSPIRVMemorySemanticsMask(const access::fence_space AccessSpace,
                             const __spv::MemorySemanticsMask LocalScopeMask =
                                 __spv::MemorySemanticsMask::WorkgroupMemory) {
-  // Huge ternary operator below is a workaround for constexpr function
-  // requirement that such function can only contain return statement and
-  // nothing more
-  //
-  // It is equivalent to the following code:
-  //
-  // uint32_t Flags =
-  //     static_cast<uint32_t>(__spv::MemorySemanticsMask::SequentiallyConsistent);
-  // switch (AccessSpace) {
-  // case access::fence_space::global_space:
-  //   Flags |=
-  //       static_cast<uint32_t>(__spv::MemorySemanticsMask::CrossWorkgroupMemory);
-  //   break;
-  // case access::fence_space::local_space:
-  //   Flags |= static_cast<uint32_t>(LocalScopeMask);
-  //   break;
-  // case access::fence_space::global_and_local:
-  // default:
-  //   Flags |= static_cast<uint32_t>(
-  //                __spv::MemorySemanticsMask::CrossWorkgroupMemory) |
-  //            static_cast<uint32_t>(LocalScopeMask);
-  //   break;
-  // }
-  // return Flags;
-
   return (AccessSpace == access::fence_space::global_space)
              ? static_cast<uint32_t>(
                    __spv::MemorySemanticsMask::SequentiallyConsistent |
@@ -60,12 +35,10 @@ getSPIRVMemorySemanticsMask(const access::fence_space AccessSpace,
              ? static_cast<uint32_t>(
                    __spv::MemorySemanticsMask::SequentiallyConsistent |
                    LocalScopeMask)
-             : /* default: (AccessSpace ==
-                  access::fence_space::global_and_local) */
-             static_cast<uint32_t>(
-                 __spv::MemorySemanticsMask::SequentiallyConsistent |
-                 __spv::MemorySemanticsMask::CrossWorkgroupMemory |
-                 LocalScopeMask);
+             : static_cast<uint32_t>(
+                   __spv::MemorySemanticsMask::SequentiallyConsistent |
+                   __spv::MemorySemanticsMask::CrossWorkgroupMemory |
+                   LocalScopeMask);
 }
 
 } // namespace detail
