+++ sycl/include/sycl/detail/sub_group_extra.hpp
