+++ sycl/test/include_deps/sycl_khr_includes_group_algorithms.hpp.cpp
@@ -16,7 +16,7 @@
 // CHECK-NEXT: detail/spirv_memory_semantics.hpp
 // CHECK-NEXT: __spirv/spirv_types.hpp
 // CHECK-NEXT: detail/defines.hpp
-// CHECK-NEXT: access/access.hpp
+// CHECK-NEXT: access/access_base.hpp
 // CHECK-NEXT: memory_enums.hpp
 // CHECK-NEXT: detail/item_base.hpp
 // CHECK-NEXT: id.hpp
@@ -31,12 +31,14 @@
 // CHECK-NEXT: detail/string.hpp
 // CHECK-NEXT: functional.hpp
 // CHECK-NEXT: group.hpp
+// CHECK-NEXT: detail/group_core.hpp
 // CHECK-NEXT: detail/assert.hpp
-// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
+// CHECK-NEXT: detail/group_extra.hpp
 // CHECK-NEXT: detail/fwd/half.hpp
-// CHECK-NEXT: detail/type_traits/integer_traits.hpp
 // CHECK-NEXT: detail/nd_loop.hpp
 // CHECK-NEXT: device_event.hpp
+// CHECK-NEXT: h_item.hpp
+// CHECK-NEXT: item.hpp
 // CHECK-NEXT: pointers.hpp
 // CHECK-NEXT: half_type.hpp
 // CHECK-NEXT: bit_cast.hpp
@@ -46,6 +48,8 @@
 // CHECK-NEXT: known_identity.hpp
 // CHECK-NEXT: aliases.hpp
 // CHECK-NEXT: detail/generic_type_traits.hpp
+// CHECK-NEXT: access/access.hpp
+// CHECK-NEXT: detail/type_traits/integer_traits.hpp
 // CHECK-NEXT: marray.hpp
 // CHECK-NEXT: detail/common.hpp
 // CHECK-NEXT: vector.hpp
@@ -54,12 +58,23 @@
 // CHECK-NEXT: detail/fwd/accessor.hpp
 // CHECK-NEXT: detail/memcpy.hpp
 // CHECK-NEXT: nd_item.hpp
-// CHECK-NEXT: item.hpp
-// CHECK-NEXT: nd_range.hpp
-// CHECK-NEXT: detail/sub_group_core.hpp
-// CHECK-NEXT: ext/oneapi/functional.hpp
+// CHECK-NEXT: detail/nd_item_core.hpp
+// CHECK-NEXT: detail/nd_item_extra.hpp
+// CHECK-NEXT: ext/oneapi/experimental/root_group.hpp
 // CHECK-NEXT: detail/spirv.hpp
 // CHECK-NEXT: __spirv/spirv_ops.hpp
 // CHECK-NEXT: multi_ptr.hpp
 // CHECK-NEXT: detail/address_space_cast.hpp
+// CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
+// CHECK-NEXT: ext/oneapi/properties/properties.hpp
+// CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
+// CHECK-NEXT: ext/oneapi/properties/property.hpp
+// CHECK-NEXT: ext/oneapi/properties/property_value.hpp
+// CHECK-NEXT: ext/oneapi/free_function_queries.hpp
+// CHECK-NEXT: detail/sub_group_core.hpp
+// CHECK-NEXT: sub_group.hpp
+// CHECK-NEXT: detail/sub_group_extra.hpp
+// CHECK-NEXT: detail/sub_group_load_store.hpp
+// CHECK-NEXT: ext/oneapi/functional.hpp
+// CHECK-NEXT: nd_range.hpp
 // CHECK-EMPTY:
