+++ sycl/test/include_deps/sycl_khr_includes_handler.hpp.cpp
@@ -10,6 +10,7 @@
 // CHECK-NEXT: feature_test.hpp
 // CHECK-NEXT: handler.hpp
 // CHECK-NEXT: access/access.hpp
+// CHECK-NEXT: access/access_base.hpp
 // CHECK-NEXT: accessor.hpp
 // CHECK-NEXT: buffer.hpp
 // CHECK-NEXT: backend_types.hpp
@@ -110,21 +111,35 @@
 // CHECK-NEXT: detail/host_profiling_info.hpp
 // CHECK-NEXT: detail/item_base.hpp
 // CHECK-NEXT: group.hpp
-// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
+// CHECK-NEXT: detail/group_core.hpp
+// CHECK-NEXT: detail/group_extra.hpp
 // CHECK-NEXT: device_event.hpp
 // CHECK-NEXT: h_item.hpp
 // CHECK-NEXT: item.hpp
 // CHECK-NEXT: kernel_handler.hpp
 // CHECK-NEXT: nd_item.hpp
-// CHECK-NEXT: nd_range.hpp
-// CHECK-NEXT: detail/sub_group_core.hpp
-// CHECK-NEXT: ext/intel/experimental/fp_control_kernel_properties.hpp
-// CHECK-NEXT: ext/oneapi/properties/property.hpp
-// CHECK-NEXT: ext/oneapi/properties/property_value.hpp
-// CHECK-NEXT: ext/intel/experimental/kernel_execution_properties.hpp
-// CHECK-NEXT: ext/oneapi/experimental/cluster_group_prop.hpp
+// CHECK-NEXT: detail/nd_item_core.hpp
+// CHECK-NEXT: detail/nd_item_extra.hpp
+// CHECK-NEXT: ext/oneapi/experimental/root_group.hpp
+// CHECK-NEXT: detail/spirv.hpp
+// CHECK-NEXT: __spirv/spirv_ops.hpp
+// CHECK-NEXT: detail/memcpy.hpp
+// CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
 // CHECK-NEXT: ext/oneapi/properties/properties.hpp
 // CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
+// CHECK-NEXT: ext/oneapi/properties/property.hpp
+// CHECK-NEXT: ext/oneapi/properties/property_value.hpp
+// CHECK-NEXT: ext/oneapi/free_function_queries.hpp
+// CHECK-NEXT: detail/sub_group_core.hpp
+// CHECK-NEXT: sub_group.hpp
+// CHECK-NEXT: detail/sub_group_extra.hpp
+// CHECK-NEXT: detail/sub_group_load_store.hpp
+// CHECK-NEXT: ext/oneapi/functional.hpp
+// CHECK-NEXT: functional.hpp
+// CHECK-NEXT: nd_range.hpp
+// CHECK-NEXT: ext/intel/experimental/fp_control_kernel_properties.hpp
+// CHECK-NEXT: ext/intel/experimental/kernel_execution_properties.hpp
+// CHECK-NEXT: ext/oneapi/experimental/cluster_group_prop.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph/command_graph.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph/common.hpp
@@ -136,7 +151,6 @@
 // CHECK-NEXT: ext/oneapi/experimental/graph/modifiable_graph.hpp
 // CHECK-NEXT: ext/oneapi/experimental/graph/dynamic.hpp
 // CHECK-NEXT: ext/oneapi/experimental/work_group_memory.hpp
-// CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
 // CHECK-NEXT: ext/oneapi/experimental/virtual_functions.hpp
 // CHECK-NEXT: ext/oneapi/kernel_properties/properties.hpp
 // CHECK-NEXT: ext/oneapi/work_group_scratch_memory.hpp
