+++ sycl/test/include_deps/sycl_khr_includes_images.hpp.cpp
@@ -11,6 +11,7 @@
 // CHECK-NEXT: accessor_image.hpp
 // CHECK-NEXT: accessor.hpp
 // CHECK-NEXT: access/access.hpp
+// CHECK-NEXT: access/access_base.hpp
 // CHECK-NEXT: buffer.hpp
 // CHECK-NEXT: backend_types.hpp
 // CHECK-NEXT: detail/array.hpp
