+++ sycl/test/include_deps/sycl_khr_includes_usm.hpp.cpp
@@ -12,13 +12,14 @@
 // CHECK-NEXT: builtins.hpp
 // CHECK-NEXT: detail/builtins/builtins.hpp
 // CHECK-NEXT: detail/fwd/multi_ptr.hpp
-// CHECK-NEXT: access/access.hpp
+// CHECK-NEXT: access/access_base.hpp
 // CHECK-NEXT: detail/loop.hpp
 // CHECK-NEXT: detail/type_traits.hpp
 // CHECK-NEXT: detail/type_traits/bool_traits.hpp
 // CHECK-NEXT: detail/type_traits/vec_marray_traits.hpp
 // CHECK-NEXT: detail/vector_convert.hpp
 // CHECK-NEXT: detail/generic_type_traits.hpp
+// CHECK-NEXT: access/access.hpp
 // CHECK-NEXT: aliases.hpp
 // CHECK-NEXT: bit_cast.hpp
 // CHECK-NEXT: detail/fwd/half.hpp
@@ -123,23 +124,36 @@
 // CHECK-NEXT: detail/item_base.hpp
 // CHECK-NEXT: detail/kernel_desc.hpp
 // CHECK-NEXT: group.hpp
-// CHECK-NEXT: detail/async_work_group_copy_ptr.hpp
+// CHECK-NEXT: detail/group_core.hpp
+// CHECK-NEXT: detail/group_extra.hpp
 // CHECK-NEXT: device_event.hpp
 // CHECK-NEXT: h_item.hpp
 // CHECK-NEXT: item.hpp
 // CHECK-NEXT: kernel_handler.hpp
 // CHECK-NEXT: nd_item.hpp
-// CHECK-NEXT: nd_range.hpp
+// CHECK-NEXT: detail/nd_item_core.hpp
+// CHECK-NEXT: detail/nd_item_extra.hpp
+// CHECK-NEXT: ext/oneapi/experimental/root_group.hpp
+// CHECK-NEXT: detail/spirv.hpp
+// CHECK-NEXT: __spirv/spirv_ops.hpp
+// CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
+// CHECK-NEXT: ext/oneapi/properties/properties.hpp
+// CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
+// CHECK-NEXT: ext/oneapi/properties/property.hpp
+// CHECK-NEXT: ext/oneapi/properties/property_value.hpp
+// CHECK-NEXT: ext/oneapi/free_function_queries.hpp
 // CHECK-NEXT: detail/sub_group_core.hpp
+// CHECK-NEXT: sub_group.hpp
+// CHECK-NEXT: detail/sub_group_extra.hpp
+// CHECK-NEXT: detail/sub_group_load_store.hpp
+// CHECK-NEXT: ext/oneapi/functional.hpp
+// CHECK-NEXT: functional.hpp
+// CHECK-NEXT: nd_range.hpp
 // CHECK-NEXT: detail/id_queries_fit_in_int.hpp
 // CHECK-NEXT: detail/nd_range_view.hpp
 // CHECK-NEXT: detail/optional.hpp
 // CHECK-NEXT: detail/range_rounding.hpp
 // CHECK-NEXT: ext/oneapi/kernel_properties/properties.hpp
-// CHECK-NEXT: ext/oneapi/properties/properties.hpp
-// CHECK-NEXT: ext/oneapi/properties/property_utils.hpp
-// CHECK-NEXT: ext/oneapi/properties/property.hpp
-// CHECK-NEXT: ext/oneapi/properties/property_value.hpp
 // CHECK-NEXT: event.hpp
 // CHECK-NEXT: exception_list.hpp
 // CHECK-NEXT: ext/oneapi/device_global/device_global.hpp
@@ -168,7 +182,6 @@
 // CHECK-NEXT: ext/intel/experimental/fp_control_kernel_properties.hpp
 // CHECK-NEXT: ext/intel/experimental/kernel_execution_properties.hpp
 // CHECK-NEXT: ext/oneapi/experimental/cluster_group_prop.hpp
-// CHECK-NEXT: ext/oneapi/experimental/use_root_sync_prop.hpp
 // CHECK-NEXT: ext/oneapi/experimental/virtual_functions.hpp
 // CHECK-NEXT: ext/oneapi/work_group_scratch_memory.hpp
 // CHECK-NEXT: detail/sycl_local_mem_builtins.hpp
